+++ packages/core/src/index.ts
@@ -4,3 +4,4 @@ export * from "./theme";
 export * from "./uno-preset";
 export * from "./mask";
 export * from "./color";
+export * from "./date-range";